pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=5.2.8",
    "django-bootstrap5>=26.1",
    "django-filter>=25.2",
    "gunicorn>=24.1.1",
    "psycopg2-binary>=2.9.11",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "rollbar>=1.3.0",
    "ruff>=0.15.2",
]


[tool.ruff]
line-length = 100
exclude = ["migrations", "venv", ".venv", "__pycache__"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N"]
ignore = ["E501"]